% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-technically-correct-data.R
\name{build_pre_technically_correct_data}
\alias{build_pre_technically_correct_data}
\title{build_pre_technically_correct_data}
\usage{
build_pre_technically_correct_data(force = FALSE)
}
\description{
build_pre_technically_correct_data
}
