% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-technically-correct-data.R
\name{copy_raw_data_weather}
\alias{copy_raw_data_weather}
\title{copy_raw_data_weather}
\usage{
copy_raw_data_weather(destination_dir = destination_dir_default,
  srource_dir = source_dir_default)
}
\description{
copy_raw_data_weather
}
