% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1a-002-charts-latitude.R
\name{generate_latitude_chart}
\alias{generate_latitude_chart}
\title{generate_latitude_chart}
\usage{
generate_latitude_chart(data_frame, x_variable_name, cluster_variable_name)
}
\arguments{
\item{data_frame}{A data_frame with "x_variable_name" and "latitude" features.}

\item{x_variable_name}{A string defining the name of the feature to be represent on the x-axis.}

\item{cluster_variable_name}{A string defining the clustering feature. Not being used at present.}
}
\description{
Generates a single scatterplot chart from a data_frame, comparing "x_variable_name" with "latitude".
}
\examples{
generate_latitude_chart(data_frame = my_data_frame, x_variable_name = "x" cluster_variable_name = "region")
}
