% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-weather-station-url.R
\name{generate_weather_station_url}
\alias{generate_weather_station_url}
\title{generate_weather_station_url}
\usage{
generate_weather_station_url(weather_station_name)
}
\description{
This function returns "Hello, world!".
}
\examples{
generate_weather_station_url()
}
