% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-technically-correct-data.R
\name{load_pre_technically_correct_data}
\alias{load_pre_technically_correct_data}
\title{load_pre_technically_correct_data}
\usage{

  load_pre_technically_correct_data(destination_dir = destination_dir_default,
  source_dir = source_dir_default)
}
\description{
load_pre_technically_correct_data
}
