% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-raw-data.R
\name{load_raw_data}
\alias{load_raw_data}
\title{load_raw_data}
\usage{
load_raw_data(weather_station_names_list = weather_station_names,
  wellbeing_destination_name = wellbeing_dest_file_name,
  wellbeing_source_name = wellbeing_source_file_name)
}
\description{
load_raw_data
}
