% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-raw-data.R
\name{load_raw_data_weather}
\alias{load_raw_data_weather}
\title{load_raw_data_weather}
\usage{

  load_raw_data_weather(weather_station_names_list = weather_station_names)
}
\description{
load_raw_data_weather
}
