% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-000-raw-data.R
\name{load_raw_data_wellbeing}
\alias{load_raw_data_wellbeing}
\title{load_raw_data_wellbeing}
\usage{

  load_raw_data_wellbeing(wellbeing_destination_name = wellbeing_dest_file_name,
  wellbeing_source_name = wellbeing_source_file_name)
}
\description{
load_raw_data_wellbeing
}
