% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-012-technically-correct-dataframe.R
\name{load_technically_correct_dataframe}
\alias{load_technically_correct_dataframe}
\title{load_technically_correct_dataframe}
\usage{

  load_technically_correct_dataframe(destination_dir = DIR_TECHNICALLY_CORRECT_WEATHER_DATAFRAME,
  source_dir = DIR_TECHNICALLY_CORRECT_WEATHER_DSV)
}
\description{
load_technically_correct_dataframe
}
