% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-011-technically-correct-dsv.R
\name{load_technically_correct_dsv}
\alias{load_technically_correct_dsv}
\title{load_technically_correct_dsv}
\usage{

  load_technically_correct_dsv(destination_dir = DIR_TECHNICALLY_CORRECT_WEATHER_DSV,
  source_dir = DIR_TECHNICALLY_CORRECT_WEATHER_TXT)
}
\description{
load_technically_correct_dsv
}
