% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-010-technically-correct-txt.R
\name{load_technically_correct_txt}
\alias{load_technically_correct_txt}
\title{load_technically_correct_txt}
\usage{

  load_technically_correct_txt(destination_dir = DIR_TECHNICALLY_CORRECT_WEATHER_TXT,
  source_dir = DIR_RAW_DATA_WEATHER)
}
\description{
load_technically_correct_txt
}
