% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-weather-station-raw-data.R
\name{load_weather_station_raw_data}
\alias{load_weather_station_raw_data}
\title{load_weather_station_raw_data}
\usage{
load_weather_station_raw_data(dest, url)
}
\description{
load_weather_station_raw_data
}
