% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-001-svc-raw-data.R
\name{question_1_001_svc_raw_data}
\alias{question_1_001_svc_raw_data}
\title{question_1_001_svc_raw_data}
\usage{
pids.wellbeing.weather::question_1_001_raw_data()
}
\arguments{
\item{weather_station_names_list}{An OPTIONAL list of weather station name strings, whose data will be downloaded.
Defaults to WEATHER_STATION_NAMES defined within constants.weather-station-names.R.}

\item{wellbeing_destination_name}{An OPTIONAL string describing the destination file name of the well-being dataset.
Defaults to WELLBEING_DEST_FILE_NAME defined within constants.wellbeing.R.}

\item{wellbeing_source_name}{An OPTIONAL string describing the source file name of the well-being dataset.
Defaults to WELLBEING_SOURCE_FILE_NAME defined within constants.wellbeing.R.}
}
\value{
A list containing 'wellbeing' and 'weather' elements, which, in turn, describe the success or failure of each file.
}
\description{
Download the raw data for both the well-being and the weather datasets.
}
