% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/question-1-002-technically-correct-weather-txt.R
\name{question_1_002_tech_correct_weather_txt}
\alias{question_1_002_tech_correct_weather_txt}
\title{question_1_002_tech_correct_weather_txt}
\usage{
pids.wellbeing.weather::question_1_002_technically_correct_weather_txt()
}
\arguments{
\item{destination_dir}{An OPTIONAL string describing the destination dir.
Defaults to DIR_TECHNICALLY_CORRECT_WEATHER_TXT defined within constants.dirs.R.}

\item{source_dir}{An OPTIONAL string describing the source dir.
Defaults to DIR_RAW_DATA_WEATHER defined within constants.dirs.R.}
}
\value{
A list containing the names of the files that have been moved.
}
\description{
Move the raw wether data into data/weather/stage-010-technically-correct-text and add a closing character.
}
