% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-003-svc-tech-weather-dsv.R
\name{question_1_003_svc_tech_weather_dsv}
\alias{question_1_003_svc_tech_weather_dsv}
\title{question_1_003_svc_tech_weather_dsv}
\usage{
pids.wellbeing.weather::question_1_003_svc_tech_weather_dsv()
}
\arguments{
\item{destination_dir}{An OPTIONAL string describing the destination dir.
Defaults to DIR_TECHNICALLY_CORRECT_WEATHER_DSV defined within constants.dirs.R.}

\item{source_dir}{An OPTIONAL string describing the source dir.
Defaults to DIR_TECHNICALLY_CORRECT_WEATHER_TXT defined within constants.dirs.R.}
}
\value{
A list describing the success or failure of each file.
}
\description{
Convert the weather files into a delimited file format with a standard number of columns.
}
