% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-004-tech-correct-weather-df.R
\name{question_1_004_tech_correct_weather_df}
\alias{question_1_004_tech_correct_weather_df}
\title{question_1_004_tech_correct_weather_df}
\usage{
pids.wellbeing.weather::question_1_004_tech_correct_weather_df()
}
\arguments{
\item{source_dir}{An OPTIONAL string describing the source dir.
Defaults to DIR_TECHNICALLY_CORRECT_WEATHER_DSV defined within constants.dirs.R.}

\item{destination_dir}{An OPTIONAL string describing the destination dir.
Defaults to DIR_TECHNICALLY_CORRECT_WEATHER_DATAFRAME defined within constants.dirs.R.}

\item{col_names}{An optional list of column names.Defaults to WEATHER_COLUMN_NAMES.}

\item{force}{An optional flag idicating whether or not the destination flags should be overwritten.}
}
\value{
A list describing the success or failure of each file.
}
\description{
Convert the weather files into a delimited file format with a standard number of columns.
}
