% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-005-svc-tech-weather-complete.R
\name{question_1_005_svc_tech_weather_complete}
\alias{question_1_005_svc_tech_weather_complete}
\title{question_1_005_svc_tech_weather_complete}
\usage{
pids.wellbeing.weather::question_1_005_svc_tech_weather_complete()
}
\arguments{
\item{source_dir}{An OPTIONAL string describing the source dir.
Defaults to DIR_TECHNICALLY_CORRECT_WEATHER_DATAFRAME defined within constants.dirs.R.}

\item{destination_dir}{An OPTIONAL string describing the destination dir.
Defaults to DIR_TECHNICALLY_CORRECT_WEATHER_COMPLETE defined within constants.dirs.R.}

\item{force}{An optional flag idicating whether or not the destination flags should be overwritten.}
}
\value{
A list describing the success or failure of each file.
}
\description{
Convert the weather files into a delimited file format with a standard number of columns.
}
