% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-007-grouped-data.R
\name{question_1_007_grouped_data}
\alias{question_1_007_grouped_data}
\title{question_1_007_grouped_data}
\usage{

  question_1_007_grouped_data(group_by_variable_name = "weather_station_name",
  scale = TRUE)
}
\arguments{
\item{group_by_variable_name}{Defines the group by variable. String. Defaults to "weather_station_name"}

\item{scale}{Whether or not the data should be scaled. Boolean. Defaults to TRUE.}
}
\description{
question_1_007_grouped_data
}
\examples{
question_1_007_grouped_data()
question_1_007_grouped_data(scale = FALSE)
qquestion_1_007_grouped_data(group_by_variable_name = "some_variable")
}
