% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-007-svc-cons-summary.R
\name{question_1_007_svc_cons_summary}
\alias{question_1_007_svc_cons_summary}
\title{question_1_007_svc_cons_summary}
\usage{
pids.wellbeing.weather::question_1_007_svc_cons_summary()
}
\arguments{
\item{df}{An OPTIONAL data frame.
Defaults to the result of load_technically_correct_data_frame().}
}
\value{
A list describing the number of 'consistently correct' rows per weather station.
}
\description{
A summary function that provides the number of 'consistently correct' rows per weather station.
}
