% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-007-svc-cons-summary.R
\name{question_1_007_svc_cons_summary_latex}
\alias{question_1_007_svc_cons_summary_latex}
\title{question_1_007_svc_cons_summary_latex}
\usage{
pids.wellbeing.weather::question_1_007_svc_cons_summary_latex()
}
\arguments{
\item{df}{An OPTIONAL data frame.
Defaults to the result of question_1_007_svc_cons_summary().}
}
\value{
A string describing the number of 'consistently correct' rows per weather station, as a Latex table.
}
\description{
A summary function that provides the number of 'consistently correct' rows per weather station as a Latex table.
}
