% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-008-grouped-data-summary.R
\name{question_1_008_grouped_data_summary}
\alias{question_1_008_grouped_data_summary}
\title{question_1_008_grouped_data_summary}
\usage{

  question_1_008_grouped_data_summary(group_by_variable_name = "weather_station_name")
}
\arguments{
\item{group_by_variable_name}{Defines the group by variable. String. Defaults to "weather_station_name"}
}
\description{
question_1_008_grouped_data_summary
}
\examples{
question_1_008_grouped_data_summary()
question_1_008_grouped_data_summary(scale = FALSE)
qquestion_1_008_grouped_data_summary(group_by_variable_name = "some_variable")
}
