% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-008-svc-cons-grouped-data.R
\name{question_1_008_svc_cons_grouped_data}
\alias{question_1_008_svc_cons_grouped_data}
\title{question_1_008_svc_cons_grouped_data}
\usage{
question_1_008_svc_cons_grouped_data()
question_1_008_svc_cons_grouped_data(scale = FALSE)
question_1_008_svc_cons_grouped_data(group_by_variable_name = "some_variable")
}
\arguments{
\item{df}{An OPTIONAL data frame.
Deafults to load_technically_correct_data_frame().}

\item{group_by_variable_name}{Defines the group by variable. String.\preformatted{   Defaults to "weather_station_name"
}}

\item{scale}{Whether or not the data should be scaled. Boolean.
Defaults to TRUE.}
}
\description{
A function that returns 'consistently correct' data grouped by weather station.
}
