% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-009-eda-charts-latitude.R
\name{question_1_009_eda_charts_latitude}
\alias{question_1_009_eda_charts_latitude}
\title{question_1_009_eda_charts_latitude}
\usage{
question_1_009_eda_charts_latitude()
}
\description{
Generate scatterplot charts comparing weather variables with latitude.
}
\examples{
question_1_009_eda_charts_latitude()
}
