% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-009-svc-cons-grouped-data-summary.R
\name{question_1_009_svc_cons_grouped_data_summary}
\alias{question_1_009_svc_cons_grouped_data_summary}
\title{question_1_009_svc_cons_grouped_data_summary}
\usage{
question_1_009_svc_cons_grouped_data_summary()
question_1_009_svc_cons_grouped_data_summary(scale = FALSE)
question_1_009_svc_cons_grouped_data_summary(group_by_variable_name = "some_variable")
}
\arguments{
\item{group_by_variable_name}{Defines the group by variable. String.
Defaults to "weather_station_name"}
}
\value{
A list containing both 'non_scaled' and 'scaled' 'consistently correct' data.
}
\description{
A function that returns list of both 'non_scaled' and 'scaled' 'consistently correct' data grouped by weather station.
}
