% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-010-eda-charts-box-whisker.R
\name{question_1_010_eda_charts_box_whisker_min_temp}
\alias{question_1_010_eda_charts_box_whisker_min_temp}
\title{question_1_010_eda_charts_box_whisker_min_temp}
\usage{
pids.wellbeing.weather::question_1_010_eda_charts_box_whisker_min_temp()
}
\value{
A ggplot2 object containing the 'box and whisker' plot.
}
\description{
A functions that returns a 'box and whisker' plot for 'min temp' per each weather station.
}
