% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-011-eda-remove-outliers.R
\name{question_1_011_eda_remove_outliers}
\alias{question_1_011_eda_remove_outliers}
\title{question_1_011_eda_remove_outliers}
\usage{
pids.wellbeing.weather::question_1_011_eda_remove_outliers()
}
\arguments{
\item{df}{An OPTIONAL data frame.
Defaults to the result of load_technically_correct_data_frame().}

\item{group_by_variable_name}{An OPTIONAL string describing the group_by feature.
Defaults to "weather_station_name".}

\item{scale}{An OPTIONAL boolean flag, indicating whther or not the results should be scaled.
Defaults to TRUE.}
}
\value{
Returns a data frame with the outliers removed.
}
\description{
A function that removed 'outliers' (greater than 1.5 * the inerquartile range)
from the 'hours_sun' and the 'rain_mm" features.
}
