% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-017-cluster-tendency.R
\name{question_1_017_cluster_tendency}
\alias{question_1_017_cluster_tendency}
\title{question_1_017_cluster_tendency}
\usage{
pids.wellbeing.weather::question_1_017_cluster_tendency()
}
\arguments{
\item{df}{An OPTIONAL data frame containing weather feature data,}

\item{show_chart}{An OPTIONAL boolean flag, incdicating whether or not a chart should be generated.}
}
\value{
A list describing the Hopkins cluster tendency
}
\description{
Calculate the Hopkins cluster tendency.
}
