% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-018-algorithm-choice.R
\name{question_1_018_algorithm_choice}
\alias{question_1_018_algorithm_choice}
\title{question_1_018_algorithm_choice}
\usage{
question_1_018_algorithm_choice(cluster_methods = c("hierarchical",
  "kmeans", "pam"), cluster_validation_type = "internal",
  num_clusters = 2:10)
}
\description{
question_1_018_algorithm_choice
}
