% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-018-vat-charts.R
\name{question_1_018_vat_charts}
\alias{question_1_018_vat_charts}
\title{question_1_018_vat_charts}
\usage{
question_1_018_vat_charts(df = question_1_010_remove_outliers())
}
\description{
question_1_018_vat_charts
}
