% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-019-prep-charts-vat.R
\name{question_1_019_prep_charts_vat}
\alias{question_1_019_prep_charts_vat}
\title{question_1_019_prep_charts_vat}
\usage{

  question_1_019_prep_charts_vat(df = question_1_011_eda_remove_outliers(scale
  = FALSE))
}
\description{
question_1_019_prep_charts_vat
}
