% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-020-prep-k-choice.R
\name{question_1_020_prep_k_choice}
\alias{question_1_020_prep_k_choice}
\title{question_1_020_prep_k_choice}
\usage{

  question_1_020_prep_k_choice(df = question_1_011_eda_remove_outliers(scale
  = TRUE), distance = "euclidean", k_max = MAX_NUM_K,
  k_min = MIN_NUM_K, method_clusters = c("centroid", "average",
  "complete", "kmeans"), method_dist = "euclidean", show_chart = TRUE)
}
\description{
question_1_020_prep_k_choice
}
