% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-021-cluster.R
\name{question_1_021_cluster}
\alias{question_1_021_cluster}
\title{question_1_021_cluster}
\usage{
question_1_021_cluster(k_max = MAX_NUM_K, metric = "euclidean",
  seed = SEED_DEFAULT)
}
\description{
question_1_021_cluster
}
