% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-021-kmediods.R
\name{question_1_021_kmediods}
\alias{question_1_021_kmediods}
\title{question_1_021_kmediods}
\usage{
question_1_021_kmediods(k_max = MAX_NUM_K,
  num_iterations = MAX_NUM_ITERATIONS, seed = SEED_DEFAULT)
}
\description{
question_1_021_kmediods
}
