% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-021-prep-algorithm-choice.R
\name{question_1_021_prep_algorithm_choice}
\alias{question_1_021_prep_algorithm_choice}
\title{question_1_021_prep_algorithm_choice}
\usage{

  question_1_021_prep_algorithm_choice(df = question_1_011_eda_remove_outliers(),
  cluster_methods = c("hierarchical", "kmeans", "pam"),
  cluster_validation_type = "internal", num_clusters = 2:3)
}
\description{
question_1_021_prep_algorithm_choice
}
