% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-022-analysis-hierarchical.R
\name{question_1_022_analysis_hierarchical}
\alias{question_1_022_analysis_hierarchical}
\title{question_1_022_analysis_hierarchical}
\usage{

  question_1_022_analysis_hierarchical(df = question_1_011_eda_remove_outliers(scale
  = TRUE), method_dist = "euclidean", method_cluster = "ward.D2")
}
\description{
question_1_022_analysis_hierarchical
}
