% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-022-prep-linkage_choice.R
\name{question_1_022_linkage_choice}
\alias{question_1_022_linkage_choice}
\title{question_1_022_linkage_choice}
\usage{

  question_1_022_linkage_choice(df = question_1_011_eda_remove_outliers(scale
  = TRUE), method_clusters = c("average", "centroid", "complete",
  "single", "ward"), method_dist = "euclidean")
}
\description{
question_1_022_linkage_choice
}
