% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-024-analysis-charts-dendrogram.R
\name{question_1_024_analysis_charts_dendrogram}
\alias{question_1_024_analysis_charts_dendrogram}
\title{qquestion_1_024_analysis_charts_dendrogram}
\usage{

  question_1_024_analysis_charts_dendrogram(hc = question_1_023_analysis_hierarchical()$hc,
  k_value = 3)
}
\description{
qquestion_1_024_analysis_charts_dendrogram
}
