% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-024-analysis-charts-sum-squares.R
\name{question_1_024_analysis_charts_sum_squares}
\alias{question_1_024_analysis_charts_sum_squares}
\title{question_1_024_analysis_charts_sum_squares}
\usage{

  question_1_024_analysis_charts_sum_squares(sum_squares = question_1_023_analysis_kmeans()$sum_squares,
  x_intercept = 3, x_intercept_line_type = 2)
}
\description{
question_1_024_analysis_charts_sum_squares
}
