% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-032-eval-charts-hier-latitude.R
\name{question_1_032_eval_charts_hierarchical_latitude}
\alias{question_1_032_eval_charts_hierarchical_latitude}
\title{question_1_032_eval_charts_hierarchical_latitude}
\usage{

  question_1_032_eval_charts_hierarchical_latitude(df = question_1_025_analysis_attach_pruned_cluster_values())
}
\description{
question_1_032_eval_charts_hierarchical_latitude
}
