% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/question-1-035-eval-charts-hier-latitude-longitude-max-temp.R
\name{question_1_035_eval_charts_hier_latitude_longitude_max_temp}
\alias{question_1_035_eval_charts_hier_latitude_longitude_max_temp}
\title{question_1_035_eval_charts_hier_latitude_longitude_max_temp}
\usage{

  question_1_035_eval_charts_hier_latitude_longitude_max_temp(df = question_1_025_analysis_attach_pruned_cluster_values(),
  x_axis_title = "Max temperature (degrees c)",
  y_axis_title = "Longitude", z_axis_title = "Latitude")
}
\description{
question_1_035_eval_charts_hier_latitude_longitude_max_temp
}
