% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1a-001-data.R
\name{question_1a_001_data}
\alias{question_1a_001_data}
\title{question_1a_001_data}
\usage{
question_1a_001_data(group_by_variable_name = "weather_station_name",
  scale = TRUE)
}
\arguments{
\item{group_by_variable_name}{Defines the group by variable. String. Defaults to "weather_station_name"}

\item{scale}{Whether or not the data should be scaled. Boolean. Defaults to TRUE.}
}
\description{
question_1a_001_data
}
\examples{
question_1a_001_data()
question_1a_001_data(scale = FALSE)
question_1a_001_data(group_by_variable_name = "some_variable")
}
