% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/question-1a-009-charts-cluster-k3-latitude-longitude.R
\name{question_1a_010_charts_box_whisker}
\alias{question_1a_010_charts_box_whisker}
\title{question_1a_010_charts_box_whisker}
\usage{
question_1a_010_charts_box_whisker()
}
\description{
question_1a_010_charts_box_whisker
}
