% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1a-010-charts-box-whisker.R
\name{question_1a_010_charts_box_whisker_hours_sun}
\alias{question_1a_010_charts_box_whisker_hours_sun}
\title{question_1a_010_charts_box_whisker_hours_sun}
\usage{
question_1a_010_charts_box_whisker_hours_sun()
}
\description{
question_1a_010_charts_box_whisker_hours_sun
}
