% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1a-010-charts-box-whisker.R
\name{question_1a_010_charts_box_whisker_rain_mm}
\alias{question_1a_010_charts_box_whisker_rain_mm}
\title{question_1a_010_charts_box_whisker_rain_mm}
\usage{
question_1a_010_charts_box_whisker_rain_mm()
}
\description{
question_1a_010_charts_box_whisker_rain_mm
}
