% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1a-010-charts-box-whisker.R
\name{question_1a_010_charts_box_whisker_temp_max}
\alias{question_1a_010_charts_box_whisker_temp_max}
\title{question_1a_010_charts_box_whisker_temp_max}
\usage{
question_1a_010_charts_box_whisker_temp_max()
}
\description{
question_1a_010_charts_box_whisker_temp_max
}
