% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1a-010-charts-box-whisker.R
\name{question_1a_010_charts_box_whisker_temp_min}
\alias{question_1a_010_charts_box_whisker_temp_min}
\title{question_1a_010_charts_box_whisker_temp_min}
\usage{
question_1a_010_charts_box_whisker_temp_min()
}
\description{
question_1a_010_charts_box_whisker_temp_min
}
