% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/question-1a-009-charts-cluster-k3-latitude-longitude.R
\name{question_1a_09_charts_cluster_k3_latitude_longitude_rain_mm}
\alias{question_1a_09_charts_cluster_k3_latitude_longitude_rain_mm}
\title{question_1a_09_charts_cluster_k3_latitude_longitude_rain_mm}
\usage{
question_1a_09_charts_cluster_k3_latitude_longitude_rain_mm()
}
\description{
question_1a_09_charts_cluster_k3_latitude_longitude_rain_mm
}
