% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1a.R
\name{question_1a_charts_latitude_comparisons}
\alias{question_1a_charts_latitude_comparisons}
\title{question_1a_charts_latitude_comparisons}
\usage{
question_1a_charts_latitude_comparisons()
}
\description{
question_1a_charts_latitude_comparisons
}
