% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1b-001-data.R
\name{question_1b_001_data}
\alias{question_1b_001_data}
\title{question_1b_001_data}
\usage{
question_1b_001_data(group_by_variable_name = "weather_station_name",
  scale = TRUE)
}
\description{
question_1b_001_data
}
