% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1b-002-charts-sum-squares.R
\name{question_1b_002_charts_sum_squares}
\alias{question_1b_002_charts_sum_squares}
\title{question_1b_002_charts_sum_squares}
\usage{
question_1b_002_charts_sum_squares(k_max = MAX_NUM_K,
  num_iterations = MAX_NUM_ITERATIONS, seed = SEED_DEFAULT)
}
\description{
question_1b_002_charts_sum_squares
}
