% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-023-charts-cluster-k2-latitude.R
\name{question_1b_005_charts_cluster_k2_latitude}
\alias{question_1b_005_charts_cluster_k2_latitude}
\title{question_1b_005_charts_cluster_k2_latitude}
\usage{
question_1b_005_charts_cluster_k2_latitude()
}
\description{
question_1b_005_charts_cluster_k2_latitude
}
