% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1b-006-charts-cluster-k3-latitude.R
\name{question_1b_006_charts_cluster_k3_latitude}
\alias{question_1b_006_charts_cluster_k3_latitude}
\title{question_1b_006_charts_cluster_k3_latitude}
\usage{
question_1b_006_charts_cluster_k3_latitude()
}
\description{
question_1b_006_charts_cluster_k3_latitude
}
