% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/question-1b-008-charts-cluster-k3-latitude-longitude.R
\name{question_1b_08_charts_cluster_k3_latitude_longitude_temp_max}
\alias{question_1b_08_charts_cluster_k3_latitude_longitude_temp_max}
\title{question_1b_08_charts_cluster_k3_latitude_longitude_temp_max}
\usage{
question_1b_08_charts_cluster_k3_latitude_longitude_temp_max()
}
\description{
question_1b_08_charts_cluster_k3_latitude_longitude_temp_max
}
