% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1e-003-charts-density-bic.R
\name{question_1c_003_charts_density_bic}
\alias{question_1c_003_charts_density_bic}
\title{question_1c_003_charts_density_bic}
\usage{
question_1c_003_charts_density_bic()
}
\description{
question_1c_003_charts_density_bic
}
