% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1d-001-hierarchical.R
\name{question_1d_001_hierarchical}
\alias{question_1d_001_hierarchical}
\title{question_1d_001_hierarchical}
\usage{
question_1d_001_hierarchical()
}
\description{
question_1d_001_hierarchical
}
