% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1d-002-hierarchical-only-temp.R
\name{question_1d_002_hierarchical_only_temp}
\alias{question_1d_002_hierarchical_only_temp}
\title{question_1d_002_hierarchical_only_temp}
\usage{
question_1d_002_hierarchical_only_temp()
}
\description{
question_1d_002_hierarchical_only_temp
}
