% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1e-001-cluster-validation.R
\name{question_1e_001_cluster_validation}
\alias{question_1e_001_cluster_validation}
\title{question_1e_001_cluster_validation}
\usage{
question_1e_001_cluster_validation(cluster_methods = c("hierarchical",
  "kmeans", "pam"), cluster_validation_type = "internal",
  num_clusters = 2:10)
}
\description{
question_1e_001_cluster_validation
}
