% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-2-001-bu-append-latitude-categories.R
\name{question_2_001_bu_append_latitude_categories}
\alias{question_2_001_bu_append_latitude_categories}
\title{question_2_001_bu_append_latitude_categories}
\usage{

  question_2_001_bu_append_latitude_categories(df = question_1_011_eda_remove_outliers(scale
  = TRUE), latitude_category_as_factor = TRUE)
}
\description{
question_2_001_bu_append_latitude_categories
}
