% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-2-002-eda-latitude-category-summary.R
\name{question_2_002_eda_latitude_category_summary}
\alias{question_2_002_eda_latitude_category_summary}
\title{question_2_002_eda_latitude_category_summary}
\usage{

  question_2_002_eda_latitude_category_summary(df = question_2_001_append_latitude_categories())
}
\description{
question_2_002_eda_latitude_category_summary
}
