% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-2-002-eda-latitude-category-summary.R
\name{question_2_002_eda_latitude_category_summary_latex}
\alias{question_2_002_eda_latitude_category_summary_latex}
\title{question_2_002_eda_latitude_category_summary_latex}
\usage{

  question_2_002_eda_latitude_category_summary_latex(res_list = question_2_002_eda_latitude_category_summary())
}
\description{
question_2_002_eda_latitude_category_summary_latex
}
