% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-2-003-eda-charts-weather-features.R
\name{question_2_003_eda_charts_weather_features}
\alias{question_2_003_eda_charts_weather_features}
\title{question_2_003_eda_charts_weather_features}
\usage{
question_2_003_eda_charts_weather_features(df = data_frame <-
  question_2_001_bu_append_latitude_categories())
}
\description{
question_2_003_eda_charts_weather_features
}
