% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/question-2-004-eda-charts-latitude_category-pairwise.R
\name{question_2_004_eda_charts_latitude_category_pairwise}
\alias{question_2_004_eda_charts_latitude_category_pairwise}
\title{question_2_004_eda_charts_latitude_category_pairwise
TODO - Fix error}
\usage{

  question_2_004_eda_charts_latitude_category_pairwise(df = question_2_001_bu_append_latitude_categories())
}
\description{
question_2_004_eda_charts_latitude_category_pairwise
TODO - Fix error
}
