% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-2-006-prep-split.R
\name{question_2_006_prep_split}
\alias{question_2_006_prep_split}
\title{question_2_006_prep_split}
\usage{

  question_2_006_prep_split(df = question_2_001_bu_append_latitude_categories(),
  num_test_rows = 5)
}
\description{
question_2_006_prep_split
}
