% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-2-007-analysis-knn.R
\name{question_2_007_analysis_knn}
\alias{question_2_007_analysis_knn}
\title{question_2_007_analysis_knn}
\usage{
question_2_007_analysis_knn(data = question_2_006_prep_split(),
  k_value = 3)
}
\description{
question_2_007_analysis_knn
}
