% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-2a-001-data.R
\name{question_2a_001_data}
\alias{question_2a_001_data}
\title{question_2a_001_data}
\usage{
question_2a_001_data(group_by_variable_name = "weather_station_name",
  scale = TRUE, latitude_category_as_factor = TRUE)
}
\description{
question_2a_001_data
}
