% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-2a-003-split.R
\name{question_2a_003_split}
\alias{question_2a_003_split}
\title{question_2a_003_split}
\usage{
question_2a_003_split(group_by_variable_name = "weather_station_name",
  scale = TRUE, num_test_rows = 5)
}
\description{
question_2a_003_split
}
