% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-2a-004-charts-correlation.R
\name{question_2a_004_charts_correlation}
\alias{question_2a_004_charts_correlation}
\title{question_2a_004_charts_correlation}
\usage{

  question_2a_004_charts_correlation(group_by_variable_name = "weather_station_name",
  scale = TRUE, latitude_category_as_factor = TRUE)
}
\description{
question_2a_004_charts_correlation
}
