% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-2-005-eda-charts-latitude-longitude.R
\name{question_2a_005_eda_charts_latitude_longitude_rain_mm}
\alias{question_2a_005_eda_charts_latitude_longitude_rain_mm}
\title{question_2_005_eda_charts_latitude_longitude_rain_mm}
\usage{

  question_2a_005_eda_charts_latitude_longitude_rain_mm(df = question_2_001_bu_append_latitude_categories())
}
\description{
question_2_005_eda_charts_latitude_longitude_rain_mm
}
