% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-2a-005-knn.R
\name{question_2a_005_knn}
\alias{question_2a_005_knn}
\title{question_2a_005_knn}
\usage{
question_2a_005_knn(group_by_variable_name = "weather_station_name",
  scale = TRUE, num_test_rows = 5, k_value = 3)
}
\description{
question_2a_005_knn
}
