% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-3-001-svc-tech-wellbeing.R
\name{question_3_001_svc_tech_wellbeing}
\alias{question_3_001_svc_tech_wellbeing}
\title{question_3_001_svc_tech_wellbeing}
\usage{

  question_3_001_svc_tech_wellbeing(source_file_path = DATA_RAW_WELLBEING_HAPPINESS_PATH,
  destination_file_path = DATA_WELLBEING_HAPPINESS_PATH, force = TRUE)
}
\description{
question_3_001_svc_tech_wellbeing
}
