% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/question-3-002-svc-cons-weather-add-boundaries.R
\name{question_3_002_svc_cons_weather_add_boundaries}
\alias{question_3_002_svc_cons_weather_add_boundaries}
\title{question_3_002_svc_cons_weather_add_boundaries}
\usage{

  question_3_002_svc_cons_weather_add_boundaries(df = question_2_001_bu_append_latitude_categories(),
  num_boundaries = 10)
}
\description{
question_3_002_svc_cons_weather_add_boundaries
}
