% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/question-3-003-svc-cons-weather-wellbeing-join.R
\name{question_3_003_svc_cons_weather_wellbeing_join}
\alias{question_3_003_svc_cons_weather_wellbeing_join}
\title{question_3_003_svc_cons_weather_wellbeing_join}
\usage{

  question_3_003_svc_cons_weather_wellbeing_join(df_happiness = load_happiness_data_frame(),
  df_weather = question_3_002_svc_cons_weather_add_boundaries())
}
\description{
question_3_003_svc_cons_weather_wellbeing_join
}
