% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/question-3-004-svc-cons-weather-wellbeing-summary.R
\name{question_3_004_svc_cons_weather_wellbeing_summary}
\alias{question_3_004_svc_cons_weather_wellbeing_summary}
\title{question_3_004_svc_cons_weather_wellbeing_summary}
\usage{

  question_3_004_svc_cons_weather_wellbeing_summary(df = question_3_003_svc_cons_weather_wellbeing_join())
}
\description{
question_3_004_svc_cons_weather_wellbeing_summary
}
