% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-3-005-eda-charts-wellbeing-by-region.R
\name{question_3_005_eda_charts_wellbeing_by_region}
\alias{question_3_005_eda_charts_wellbeing_by_region}
\title{question_3_005_eda_charts_wellbeing_by_region}
\usage{

  question_3_005_eda_charts_wellbeing_by_region(df = question_3_003_svc_cons_weather_wellbeing_join())
}
\description{
question_3_005_eda_charts_wellbeing_by_region
}
