% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-3-007-analysis-regression.R
\name{question_3_007_analysis_regression}
\alias{question_3_007_analysis_regression}
\title{question_3_007_analysis_regression}
\usage{

  question_3_007_analysis_regression(df_training = question_3_003_svc_cons_weather_wellbeing_join())
}
\description{
question_3_007_analysis_regression
}
