% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-3-008-analysis-regression-test.R
\name{question_3_008_analysis_regression_test}
\alias{question_3_008_analysis_regression_test}
\title{question_3_008_analysis_regression_test}
\usage{

  question_3_008_analysis_regression_test(df_test = question_3_002_svc_cons_weather_add_boundaries(),
  model = question_3_007_analysis_regression())
}
\description{
question_3_008_analysis_regression_test
}
