% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-3-009-eval-regression-summary.R
\name{question_3_009_eval_regression_summary}
\alias{question_3_009_eval_regression_summary}
\title{question_3_009_eval_regression_summary}
\usage{

  question_3_009_eval_regression_summary(df = question_3_007_analysis_regression())
}
\description{
question_3_009_eval_regression_summary
}
