% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-3-009-eval-regression-summary.R
\name{question_3_009_eval_regression_summary_latex}
\alias{question_3_009_eval_regression_summary_latex}
\title{question_3_009_eval_regression_summary_latex}
\usage{

  question_3_009_eval_regression_summary_latex(df = question_3_009_eval_regression_summary())
}
\description{
question_3_009_eval_regression_summary_latex
}
