% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-3a-001-data-weather.R
\name{question_3a_001_data_weather}
\alias{question_3a_001_data_weather}
\title{question_3a_001_data_weather}
\usage{
question_3a_001_data_weather(num_boundaries = 10)
}
\description{
question_3a_001_data_weather
}
