% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-3a-004-regression-model.R
\name{question_3a_004_regression_model}
\alias{question_3a_004_regression_model}
\title{question_3a_004_regression_model}
\usage{
question_3a_004_regression_model()
}
\description{
question_3a_004_regression_model
}
