% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-3a-005-regression-summary.R
\name{question_3a_005_regression_summary}
\alias{question_3a_005_regression_summary}
\title{question_3a_005_regression_summary}
\usage{
question_3a_005_regression_summary()
}
\description{
question_3a_005_regression_summary
}
