% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-3a-006-regression-test.R
\name{question_3a_006_regression_test}
\alias{question_3a_006_regression_test}
\title{question_3a_006_regression_test}
\usage{
question_3a_006_regression_test()
}
\description{
question_3a_006_regression_test
}
