% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-010-technically-correct-txt.R
\name{stage_010}
\alias{stage_010}
\title{stage_010}
\usage{
stage_010(destination_dir = DIR_TECHNICALLY_CORRECT_WEATHER_TXT,
  source_dir = DIR_RAW_DATA_WEATHER)
}
\description{
stage_010
}
