% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-011-technically-correct-weather-dsv.R
\name{stage_011}
\alias{stage_011}
\title{stage_011}
\usage{
stage_011(destination_dir = DIR_TECHNICALLY_CORRECT_WEATHER_DSV,
  source_dir = DIR_TECHNICALLY_CORRECT_WEATHER_TXT)
}
\description{
stage_011
}
