% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/stage-011-technically-correct-weather-dsv-revised.R
\name{stage_011_revised}
\alias{stage_011_revised}
\title{stage_011_revised}
\usage{
stage_011_revised(destination_dir = DIR_TECHNICALLY_CORRECT_WEATHER_DSV,
  source_dir = DIR_TECHNICALLY_CORRECT_WEATHER_TXT)
}
\description{
stage_011_revised
}
