% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-012-technically-correct-dataframe.R
\name{stage_012}
\alias{stage_012}
\title{stage_012}
\usage{
stage_012(source_dir = DIR_TECHNICALLY_CORRECT_WEATHER_DSV,
  destination_dir = DIR_TECHNICALLY_CORRECT_WEATHER_DATAFRAME)
}
\description{
stage_012
}
