% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/stage-014-technically-correct-weather-single-dataframe.R
\name{stage_014}
\alias{stage_014}
\title{stage_014}
\usage{
stage_014(source_dir = DIR_TECHNICALLY_CORRECT_WEATHER_COMPLETE,
  destination_dir = DIR_TECHNICALLY_CORRECT_ALL, force = TRUE)
}
\description{
stage_014
}
