% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-015-technically-correct-happiness.R
\name{stage_015}
\alias{stage_015}
\title{stage_015}
\usage{
stage_015(source_file_path = DATA_RAW_WELLBEING_HAPPINESS_PATH,
  destination_file_path = DATA_WELLBEING_HAPPINESS_PATH, force = TRUE)
}
\description{
stage_015
}
