% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather-station-names.R
\docType{data}
\name{weather_station_names}
\alias{weather_station_names}
\title{weather_station_names}
\format{An object of class \code{list} of length 36.}
\usage{
weather_station_names
}
\description{
weather_station_names
}
\keyword{datasets}
