% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-1-017-cluster-tendency.R
\name{question_1_017_cluster_tendency}
\alias{question_1_017_cluster_tendency}
\title{question_1_017_cluster_tendency}
\usage{
pids.wellbeing.weather::question_1_017_cluster_tendency()
}
\arguments{
\item{group_by_variable_name}{An OPTIONAL string describing the group_by variable. Defaults to "weather_station_name".}

\item{scale}{An OPTIONAL boolean indicating whether or not the data should be scaled.}
}
\value{
A list describing the Hopkins cluster tendency
}
\description{
Calculate the Hopkins cluster tendency.
}
