% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-000-raw-data.R
\name{stage_000}
\alias{stage_000}
\title{stage_000}
\usage{
stage_000(weather_station_names_list = WEATHER_STATION_NAMES,
  wellbeing_destination_name = wellbeing_dest_file_name,
  wellbeing_source_name = wellbeing_source_file_name)
}
\description{
stage_000
}
